Dev to Main

+++ app.js
@@ -5,8 +5,6 @@ const jwt = require('jsonwebtoken');
 
 require('dotenv').config();
 
-// const jwtSecret = process.env.JWT_SECRET;
-
 const basePath = "/api/v1";
 
 const Film = require('./models/film_model');
@@ -32,29 +30,69 @@ app.post(basePath + "/login", async (req, res) => {
     password: req.body.password
   }
 
-  try {
-    await User.countDocuments({ username: user.username, password: user.password }, function (e, count) {
-      if (count > 0) {
-        console.log(":: User " + user.username + " has access, issuing token ::");
-        const jwtSecret = utils.getJwtSecret(user.password);
-        jwt.sign({ user }, jwtSecret, (err, token) => {
-          res.json({
-            status: 200,
-            token
-          })
-        })
-      } else {
-        console.log(":: User " + user.username + " not found ::");
-        utils.returnError(res, 403, "User " + user.username + " does not exist or password is incorrect");
-      }
-    }).clone()
-      .catch(e => { 
-        console.log(e);
-        utils.returnError(res, 400, e.message);
-      });
-  } catch (e) {
-    console.log(e);
-    utils.returnError(res, 400, e.message);
+  function signTokenForUser(status) {
+    const jwtSecret = utils.getJwtSecret(user.password);
+    jwt.sign({ user }, jwtSecret, (err, token) => {
+      res.json({
+        status: status,
+        token
+      })
+    })
+  }
+
+  const isNew = req.query.isNew;
+
+  if (!isNew) {
+
+    try {
+      await User.countDocuments({ username: user.username, password: user.password }, function (e, count) {
+        if (count > 0) {
+          console.log(":: User " + user.username + " has access, issuing token ::");
+          signTokenForUser(200);
+        } else {
+          console.log(":: User " + user.username + " not found ::");
+          utils.returnError(res, 403, "User " + user.username + " does not exist or password is incorrect");
+        }
+      }).clone()
+        .catch(e => { 
+          console.log(e);
+          utils.returnError(res, 400, e.message);
+        });
+    } catch (e) {
+      console.log(e);
+      utils.returnError(res, 400, e.message);
+    }
+  } else {
+    try {
+      await User.countDocuments({ username: user.username }, function (e, count) {
+        if (count > 0) { // if user exists, return error
+          console.log(":: User " + user.username + " already exists ::");
+          utils.returnError(res, 400, "User " + user.username + " already exists");
+        } else { // if user does not exist, create user
+          console.log(":: User " + user.username + " does not exist, creating ::");
+          const newUser = new User({ // create new user object
+            username: user.username,
+            password: user.password
+          });
+          newUser.save() // save user
+            .then(() => {
+              console.log(":: User " + user.username + " created ::");
+              signTokenForUser(201);
+            })
+            .catch(e => {
+              console.log(e);
+              utils.returnError(res, 400, e.message);
+            });
+        }
+      }).clone()
+        .catch(e => {
+          console.log(e);
+          utils.returnError(res, 400, e.message);
+        });
+    } catch (e) {
+      console.log(e);
+      utils.returnError(res, 400, e.message);
+    }
   }
   
 })
@@ -89,6 +127,33 @@ app.get(basePath + "/films", utils.verifyToken, async (req, res) => {
   
 });
 
+// GET /films/:id - to get a film by its id
+app.get(basePath + "/films/:id", utils.verifyToken, async (req, res) => {
+  const id = req.params.id;
+  let addedBy = utils.decodeToken(req.headers['authorization']).user.username;
+
+  if (utils.checkIdLength(id)) { // if id is a valid mongoose id
+    try {
+      // const films = await Film.find({}); // get all the films that the user added to DB
+      var film = await Film.findOne({_id: id, addedBy: addedBy});
+
+      if (film) {
+        console.log(":: Returning the film :: ID " + id);
+        res.json(film);
+      } else {
+        console.log(":: Film not found for user :: ID " + id);
+        utils.returnError(res, 404, "Film not found for user");
+      }
+    } catch (e) {
+      console.log(e);
+      utils.returnError(res, 400, e.message);
+    }
+  } else { // if id is not valid
+    console.log(":: Invalid ID ::");
+    utils.returnError(res, 400, "Invalid ID");
+  }       
+});
+
 // POST /films - To add a film to the list
 app.post(basePath + "/films", utils.verifyToken, async (req, res) => {
   let addedBy = utils.decodeToken(req.headers['authorization']).user.username;
@@ -99,18 +164,22 @@ app.post(basePath + "/films", utils.verifyToken, async (req, res) => {
     var film = new Film({
       name: req.body.name,
       rating: req.body.rating,
+      releaseYear: req.body.releaseYear,
+      posterUrl: req.body.posterUrl,
+      notes: req.body.notes,
       addedBy: addedBy,
       insertedOn: today
     })
 
     await film.save(); // save the film data to DB
 
-    console.log(":: Film '" + film.name + "' was created ::");
+    console.log(":: Film '" + film.name + "' was created :: ID: " + film._id);
 
     res.status(201);
     res.json({
       status: 201,
-      message: "Film was successfully added."
+      message: "Film was successfully added",
+      id: film._id
     });
   } catch(e) { // catch any errors thrown
     console.log(e);
@@ -123,42 +192,51 @@ app.put(basePath + "/films/:id", utils.verifyToken, async (req, res) => {
   const id = req.params.id;
   const addedBy = utils.decodeToken(req.headers['authorization']).user.username;
 
-  // var updatedFilm = new Film({
-  //   name: req.body.name,
-  //   rating: req.body.rating
-  // })
-
-  const name = req.body.name;
-  const rating = req.body.rating;
-
-  if (name || rating) {
-    try {
-    
-      // var film = await Film.findOneAndUpdate({_id: id, addedBy: addedBy}, updatedFilm, { new: true })
-      var film = await Film.findOne({_id: id, addedBy: addedBy});
-  
-      if (film) {
-        if (name) film.name = name;
-        if (rating) film.rating = rating;
-        await film.save();
-  
-        console.log(":: Film with ID " + id + " was updated ::")
-        res.json({
-          status: 200,
-          message: "Film with ID " + id + " was updated"
-        })
-      } else {
-        utils.returnError(res, 404, "Film not found for user");
-      }
-  
-    } catch (e) {
-      console.log(e);
-      utils.returnError(res, 400, e.message);
-    }
-  } else {
-    utils.returnError(res, 400, "Please supply a name or rating value to update for the film.")
+  var updatedData = {
+    name: req.body.name,
+    rating: req.body.rating,
+    releaseYear: req.body.releaseYear,
+    posterUrl: req.body.posterUrl,
+    notes: req.body.notes,
+    updatedOn: new Date()
   }
 
+  if (utils.checkIdLength(id)) { // if id is a valid mongoose id
+    if (updatedData.name || updatedData.rating || updatedData.releaseYear || updatedData.posterUrl || updatedData.notes) { // if there is data to update
+      try {
+        var film = await Film.findOne({_id: id, addedBy: addedBy});
+    
+        if (film) {
+          if (updatedData.name) film.name = updatedData.name;
+          if (updatedData.rating) film.rating = updatedData.rating;
+          if (updatedData.releaseYear) film.releaseYear = updatedData.releaseYear;
+          if (updatedData.posterUrl) film.posterUrl = updatedData.posterUrl;
+          if (updatedData.notes) film.notes = updatedData.notes;
+          film.updatedOn = updatedData.updatedOn;
+
+          await film.save();
+    
+          console.log(":: Film with ID " + id + " was updated ::")
+          res.json({
+            status: 200,
+            message: "Film was successfully updated",
+            id: film._id
+          })
+        } else {
+          utils.returnError(res, 404, "Film not found for user");
+        }
+    
+      } catch (e) {
+        console.log(e);
+        utils.returnError(res, 400, e.message);
+      }
+    } else {
+      utils.returnError(res, 400, "Please supply at least one updated field for the film.")
+    }
+  } else { // if id is not valid
+    console.log(":: Invalid ID ::");
+    utils.returnError(res, 400, "Invalid ID");
+  }
 })
 
 // DELETE /films - to clear the entire film list
@@ -188,61 +266,26 @@ app.delete(basePath + "/films/:id", utils.verifyToken, async (req, res) => {
   let id = req.params.id;
   let addedBy = utils.decodeToken(req.headers['authorization']).user.username;
 
-  try {
-    let film = await Film.findOneAndDelete({_id: id, addedBy: addedBy});
-    if (film) {
-      console.log(":: Film with ID " + id + " was removed ::");
-      res.json({
-        status: 200,
-        message: "Film was removed"
-      });
-    } else {
-      utils.returnError(res, 404, "Film not found for user");
-    } 
-  } catch(e) {
-    console.log(e);
-    utils.returnError(res, 400, e.message);
+  if (utils.checkIdLength(id)) { // if id is a valid mongoose id
+    try {
+      let film = await Film.findOneAndDelete({_id: id, addedBy: addedBy});
+      if (film) {
+        console.log(":: Film with ID " + id + " was removed ::");
+        res.json({
+          status: 200,
+          message: "Film was removed"
+        });
+      } else {
+        utils.returnError(res, 404, "Film not found for user");
+      } 
+    } catch(e) {
+      console.log(e);
+      utils.returnError(res, 400, e.message);
+    }
+  } else {
+    utils.returnError(res, 400, "Invalid ID");  // if id is not valid
   }
   
 })
-
-
-// reusable function to return errors
-// function utils.returnError(res, stat, msg) {
-//     res.status(stat);
-//     res.json({
-//       status: stat,
-//       message: msg
-//     })
-// }
-
-// // function to validate the jwt token passed in on certain endpoints
-// function verifyToken(req, res, next) {
-//   const bearerHeader = req.headers['authorization'];
-//   if (typeof bearerHeader !== 'undefined') {
-//     const bearerToken = bearerHeader.split(' ')[1];
-//     jwt.verify(bearerToken, jwtSecret, (err, authData) => {
-//       if (err) {
-//         utils.returnError(res, 403, err.message);
-//       } else {
-//         next();
-//       }
-//     })
-//   } else {
-//     utils.returnError(res, 403, 'Token is invalid')
-//   }
-// }
-
-// // function to check if user is associated with the film
-// function checkfilmAddedByUser(req) {
-//   let id = req.params.id;
-//   let addedBy = jwt.decode(req.headers['authorization'].split(' ')[1]).user.username;
-
-//   let film = await Film.find({_id: id, addedBy: addedBy});
-
-//   if (film === 0) return false
-//   else return true
-// }
-
 
 module.exports = app;
